+++ tests/codegen/cpp/test_command.cpp
@@ -35,6 +35,7 @@
 #include "command/test_ack_packet11.hpp"
 #include "command/test_command_packet_wif110.hpp"
 #include "command/test_ack_packet_wif110.hpp"
+#include "command/difi1p2.hpp"
 
 #include <vector>
 #include <bytes.hpp>
@@ -113,7 +114,8 @@ TEST_CASE("Section 8.2", "[command_packet][8.2]")
             // auto CONTROLLEE_FORMAT_128 = uint8_t{0x1 << 6 }; // Bit 30 in CAM
             // auto CONTROLLER_ENABLED = uint8_t{0x1 << 5 }; // Bit 29 in CAM
             // auto CONTROLLER_FORMAT_128 = uint8_t{0x1 << 4 }; // Bit 28 in CAM
-            bytes CAM_BE = { 0xF0, 0, 0, 0 };
+            // auto ACTION_MODE_EXECUTE = uint8_t{0x1}; // Bit 24 in CAM
+            bytes CAM_BE = { 0xF1, 0, 0, 0 };
             auto MESSAGE_ID_BE = parse_32(MESSAGE_ID);
             auto data = packet_in.data();
             auto* check_ptr = data.data();
@@ -162,7 +164,8 @@ TEST_CASE("Section 8.2", "[command_packet][8.2]")
 
         bytes HEADER_BE { 0x60, 0x00, 0x00, static_cast<uint8_t>(PACKET_SIZE/4) };
         auto STREAM_ID_BE = parse_32(STREAM_ID);
-        bytes CAM_BE = { 0, 0, 0, 0 };
+	// auto ACTION_MODE_EXECUTE = uint8_t{0x1}; // Bit 24 in CAM
+        bytes CAM_BE = { 0x01, 0, 0, 0 };
         auto MESSAGE_ID_BE = parse_32(MESSAGE_ID);
         auto data = packet_in.data();
         auto* check_ptr = data.data();
@@ -807,5 +810,52 @@ TEST_CASE("Section 8.4", "[acknowledge_packet_subtype][8.4]")
             CHECK(packet_out.scheduled_or_executed() == true);
             CHECK(packet_out.cam().scheduled_or_executed() == true);
         }
+    }
+}
+
+TEST_CASE("DIFI Section 4.3", "[command_packet_classes][4.3]")
+{
+    SECTION("4.3")
+    {
+        size_t PACKET_SIZE = HEADER_BYTES + 
+                            STREAM_ID_BYTES + 
+                            CAM_BYTES + 
+                            MESSAGE_ID_BYTES;
+        SECTION("Buffer Size")
+        {
+            const size_t BUFFER_SIZE_BYTES = 12;
+            PACKET_SIZE += CIF0_BYTES + CIF1_BYTES + BUFFER_SIZE_BYTES;
+
+            Difi1p2 packet_in;
+            CHECK(packet_in.size() == PACKET_SIZE);
+
+            uint64_t BUFFER_SIZE = 0xBA5EBA11BA5EBA11;
+	    uint16_t LEVEL = 0x0BEE;
+	    bool ONE = true;
+	    packet_in.buffer_size().buffer_size(BUFFER_SIZE);
+	    packet_in.buffer_size().level(LEVEL);
+	    packet_in.buffer_size().overflow(ONE);
+	    packet_in.buffer_size().nearly_full(ONE);
+	    packet_in.buffer_size().nearly_empty(ONE);
+	    packet_in.buffer_size().underflow(ONE);
+
+	    auto data = packet_in.data();
+	    CHECK(data.size() == PACKET_SIZE);
+            // Check bytes
+            auto* check_ptr = data.data();
+            check_ptr += HEADER_BYTES + STREAM_ID_BYTES + CAM_BYTES + MESSAGE_ID_BYTES + CIF0_BYTES + CIF1_BYTES;
+            bytes packed_buffer_size(check_ptr, check_ptr + 12);
+	    CHECK(packed_buffer_size == bytes{ 0xBA, 0x5E, 0xBA, 0x11, 0xBA, 0x5E, 0xBA, 0x11, 0x00, 0x00, 0xBE, 0xEF });
+            
+            Difi1p2 packet_out(data);
+            CHECK(packet_out.size() == PACKET_SIZE);
+	    CHECK(packet_out.buffer_size().buffer_size() == BUFFER_SIZE);
+	    CHECK(packet_out.buffer_size().level() == LEVEL);
+	    CHECK(packet_out.buffer_size().overflow() == ONE);
+	    CHECK(packet_out.buffer_size().nearly_full() == ONE);
+	    CHECK(packet_out.buffer_size().nearly_empty() == ONE);
+	    CHECK(packet_out.buffer_size().underflow() == ONE);
+
+	}
     }
 }