+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.14] - 2024-11-06
+### Added
+  - YAML tags and namespace to support DIFI 1.2 deviations to reference level and buffer size context fields
+  - Support for packet types in multiple information classes
+  - 'any' type for tsm fields
+    - Adds additional getter/setter functions for tsm
+    - Removes explicit matching on tsm
+### Changed
+- Make Class ID field `pad_bits` runtime dynamic
+  - Do not check value in match function
+  - Add public getter/setter functions
+- Removed setter for control packet CAM action mode field
+### Fixed
+- Vrtpktgen bug when no acknowledgement packets are requested
 ## [0.7.13] - 2024-09-19
 ### Fixed
 - Threshold data model field variable names